+++ src/Components/Main/About/About.scss
@@ -1,5 +1,4 @@
-@use '../../../assets/scss/variables';
-
+@use '../../../styles/variables';
 .about {
   margin: 0 70px 70px;
