+++ src/Components/Main/Button/Button.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/scss/variables';
+@use '../../../styles/variables';
 
 .courses__btn {
   display: flex;
