+++ src/Components/Main/Card/Card.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/scss/variables';
+@use '../../../styles/variables';
 
 .card {
   max-width: 700px;
