+++ src/Components/Main/Courses/Courses.scss
@@ -1,4 +1,5 @@
-@use '../../../assets/scss/variables';
+@use '../../../styles/variables';
+
 .courses {
   margin: 0 70px 70px;
