+++ src/Components/Main/Hero/Hero.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/scss/variables';
+@use '../../../styles/variables';
 
 .hero {
   display: flex;
