+++ src/Components/Main/Layouts/Layout.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/scss/variables';
+@use '../../../styles/variables';
 
 .layout {
   clear: right;
