+++ src/styles/global.scss
@@ -1,4 +1,4 @@
-@use './assets/scss/variables';
+@use './variables';
 @import url('https://fonts.googleapis.com/css2?family=Cinzel:wght@400;500;700&family=Roboto:wght@100;300;400;500;700&family=Russo+One&family=Zen+Kurenaido&display=swap');
 
 * {
